+++ packages/gasket-plugin-intl/lib/apm-transaction.js
@@ -1,0 +1,3 @@
+module.exports = (_gasket, transaction, { res }) => {
+  transaction.setLabel('locale', res.locals.gasketData.locale);
+};