+++ packages/gasket-plugin-intl/lib/index.js
@@ -5,6 +5,7 @@ const init = require('./init');
 const middleware = require('./middleware');
 const serve = require('./serve');
 const serviceWorkerCacheKey = require('./service-worker-cache-key');
+const apmTransaction = require('./apm-transaction');
 const workbox = require('./workbox');
 const buildManifest = require('./build-manifest');
 const buildModules = require('./build-modules');
@@ -58,7 +59,13 @@ module.exports = {
     },
     express: serve,
     fastify: serve,
-    middleware,
+    middleware: {
+      timing: {
+        before: ['@gasket/plugin-elastic-apm'],
+        handler: middleware
+      }
+    },
+    apmTransaction,
     workbox,
     serviceWorkerCacheKey,
     metadata(gasket, meta) {