+++ packages/gasket-plugin-nextjs/lib/index.js
@@ -4,6 +4,8 @@ const { name, devDependencies } = require('../package');
 const { createConfig } = require('./config');
 const { pluginIdentifier } = require('@gasket/resolve');
 const { setupNextApp } = require('./setup-next-app');
+const getNextRoute = require('./next-route');
+const apmTransaction = require('./apm-transaction');
 
 const isDefined = (o) => typeof o !== 'undefined';
 
@@ -93,6 +95,17 @@ module.exports = {
         }
       }
     },
+    middleware: {
+      timing: {
+        before: ['@gasket/plugin-elastic-apm']
+      },
+      handler: (gasket) => [
+        (req, _res, next) => {
+          req.getNextRoute = () => getNextRoute(gasket, req);
+          next();
+        }
+      ]
+    },
     express: {
       timing: {
         last: true
@@ -164,6 +177,7 @@ module.exports = {
         return app;
       }
     },
+    apmTransaction,
     build: async function build(gasket) {
       const { command } = gasket;
       // Don't do a build, use dev server for local